Move into CLI folder and store additional info in saved file
